requirements.txt
python-dotenv==1.0.1
requests==2.32.3
tenacity==8.5.0
pandas==2.2.2
numpy==1.26.4
gspread==6.1.2
google-auth==2.34.0
rapidfuzz==3.9.6
openai==1.51.2